requirements.txt
requests
beautifulsoup4
lxml
aiohttp
asyncio
sqlalchemy
psycopg2
loguru
python-dateutil
PyYAML
pytest
pytest-asyncio
fake-useragent
curl-cffi
playwright
python-dotenv